requirements.txt
streamlit==1.10.0
pdfplumber==0.6.0
pyspellchecker==0.6.3
